feat: update DeleteReview to enforce ownership check and add corresponding tests

+++ app/Actions/Review/DeleteReview.php
@@ -3,11 +3,16 @@
 namespace App\Actions\Review;
 
 use App\Models\Review;
+use App\Models\User;
 
 class DeleteReview
 {
-    public function handle(Review $review): void
+    public function handle(User $user, Review $review): void
     {
+        if(! $review->isOwnedBy($user)) {
+            throw new \DomainException('You cannnot delete this review.');
+        }
+
         $review->delete();
     }
 }
